chore: move metadata logic to new `server/` directory
